refactor: class locations and packages
