+++ src/main/kotlin/com/ingsis/snippets/snippet/SnippetController.kt
@@ -1,7 +1,5 @@
-package com.ingsis.snippets.snippet.routes
+package com.ingsis.snippets.snippet
 
-import com.ingsis.snippets.snippet.SnippetService
-import com.ingsis.snippets.snippet.model.Snippet
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
