+++ src/main/kotlin/com/ingsis/snippets/snippet/SnippetRepository.kt
@@ -1,6 +1,5 @@
 package com.ingsis.snippets.snippet
 
-import com.ingsis.snippets.snippet.model.Snippet
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
